+++ lib/ooxml_parser/xlsx_parser/xlsx_data/view_model/workbook/pivot_table_definition/pivot_fields/pivot_field.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'pivot_field/items'
+
 module OoxmlParser
   # Class for parsing <pivotField> tag
   class PivotField < OOXMLDocumentObject
@@ -7,6 +9,8 @@ module OoxmlParser
     attr_reader :axis
     # @return [True, False] should show all
     attr_reader :show_all
+    # @return [Items] contain item
+    attr_reader :items
 
     # Parse `<pivotField>` tag
     # # @param [Nokogiri::XML:Element] node with PivotField data
@@ -18,6 +22,13 @@ module OoxmlParser
           @axis = value.value.to_s
         when 'showAll'
           @show_all = attribute_enabled?(value)
+        end
+      end
+
+      node.xpath('*').each do |node_child|
+        case node_child.name
+        when 'items'
+          @items = Items.new(parent: self).parse(node_child)
         end
       end
       self